+++ src/app/app.module.ts
@@ -6,14 +6,14 @@ import { NgxDatatableModule } from '@swimlane/ngx-datatable';
 
 import { AppComponent } from './app.component';
 import { CounterComponent } from './counter.component';
-// import { NgxUIModule } from '@swimlane/ngx-ui';
+import { MomentModule } from 'ngx-moment';
 import { NgxUIModule } from '../../projects/swimlane/ngx-ui/src/public_api';
 
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [AppComponent, CounterComponent],
-  imports: [BrowserModule, HttpClientModule, NgxUIModule, FormsModule, NgxDatatableModule, BrowserAnimationsModule],
+  imports: [BrowserModule, HttpClientModule, MomentModule, NgxUIModule, FormsModule, NgxDatatableModule, BrowserAnimationsModule],
   bootstrap: [AppComponent]
 })
 export class AppModule {}